fix: circular import of _check_dims etc. (now sits in core.utils)
